requirements.txt
Cython==3.0.0a9
PyOpenGL
PyOpenGL_accelerate